pyproject.toml
[project]
name = "bio-face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepface>=0.0.97",
    "fastapi[standard]>=0.128.0",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "pypdf>=6.6.0",
    "reportlab>=4.4.9",
    "tf-keras>=2.20.1",
    "torch>=2.9.1",
    "uvicorn>=0.40.0",
]
